Moved logical label code to useKanban and removed unused dependancies
